Refactor: Move reputation list out to own MCV package
