+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/user/reputation/ReputationModel.java
@@ -18,20 +18,10 @@
 package bisq.desktop.main.content.user.reputation;
 
 import bisq.desktop.common.view.Model;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.collections.transformation.FilteredList;
-import javafx.collections.transformation.SortedList;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Getter
 public class ReputationModel implements Model {
-    private final ObservableList<ReputationView.ListItem> listItems = FXCollections.observableArrayList();
-    private final FilteredList<ReputationView.ListItem> filteredList = new FilteredList<>(listItems);
-    private final SortedList<ReputationView.ListItem> sortedList = new SortedList<>(filteredList);
-    private final StringProperty userProfileIdOfScoreUpdate = new SimpleStringProperty();
 }
